fix(backend): make ChatModel temperature a float instead of int (#784)
Ref: #783

+++ python/beeai_framework/backend/types.py
@@ -30,7 +30,7 @@ class ChatModelParameters(BaseModel):
     max_tokens: int | None = None
     top_p: int | None = None
     frequency_penalty: int | None = None
-    temperature: int = 0
+    temperature: float = 0
     top_k: int | None = None
     n: int | None = None
     presence_penalty: int | None = None
